requirements.txt
pandas==2.3.2
duckdb==1.3.2
dbt-core==1.10.10
dbt-duckdb==1.9.4
langchain==0.3.21
langchain-ollama==0.3.0
langchain-google-genai==2.1.10
python-dotenv==1.1.0
rich==13.9.4
streamlit==1.49.1
watchdog==6.0.0
